+++ src/main/kotlin/com/siliconvalleyoffice/git4jira/constant/StringConstant.kt
@@ -1,4 +1,4 @@
-package com.siliconvalleyoffice.git4jira.app
+package com.siliconvalleyoffice.git4jira.constant
 
 import java.util.*
