+++ src/main/kotlin/com/siliconvalleyoffice/git4jira/service/json/JsonFileService.kt
@@ -2,14 +2,14 @@ package com.siliconvalleyoffice.git4jira.service.json
 
 import com.google.gson.Gson
 import com.siliconvalleyoffice.git4jira.util.CONFIG
-import com.siliconvalleyoffice.git4jira.app.Git4JiraApp
+import com.siliconvalleyoffice.git4jira.Git4JiraApp
 import com.siliconvalleyoffice.git4jira.util.PROJECT_PROFILE
 import com.siliconvalleyoffice.git4jira.model.Configuration
 import com.siliconvalleyoffice.git4jira.model.Project
 import com.siliconvalleyoffice.git4jira.model.ProjectProfileData
 import com.siliconvalleyoffice.git4jira.service.Service
 
-class JsonFilesService(val gson: Gson) : Service.JsonFiles {
+class JsonFileService(val gson: Gson) : Service.JsonFiles {
 
     override lateinit var configuration: Configuration
     override lateinit var projectProfileData: ProjectProfileData
